色々修正

+++ server.go
@@ -9,34 +9,44 @@ import (
 	"github.com/labstack/echo"
 )
 
-type HTTPMethod int
-
+// 各メソッドの定義
 const (
-	GET HTTPMethod = iota
+	GET int = iota
 	POST
 	PUT
 	DELETE
 )
 
+// Route エンドポイント管理
 type Route struct {
-	method  HTTPMethod
+	method  int
 	path    string
 	handler echo.HandlerFunc
 }
 
+// APIのルーティングの指定
 var routeAPI = []Route{
 	{GET, "/api/game", view.GetGame},
 }
 
-func startServer(host string) {
+// startServer サーバーを起動する
+func startServer(sock bool) {
 	e := echo.New()
-
-	os.Remove("/tmp/echo.sock")
-	l, err := net.Listen("unix", "/tmp/echo.sock")
-	if err != nil {
-		e.Logger.Fatal(err)
+	// unix sockでlistenするか
+	if sock {
+		os.Remove("/tmp/echo.sock")
+		l, err := net.Listen("unix", "/tmp/echo.sock")
+		if err != nil {
+			e.Logger.Fatal(err)
+		}
+		e.Listener = l
+	} else {
+		l, err := net.Listen("tcp", ":8080")
+		if err != nil {
+			e.Logger.Fatal(l)
+		}
+		e.Listener = l
 	}
-	e.Listener = l
 
 	for _, route := range routeAPI {
 		switch route.method {
